VISTA-2605 correction to domain definition

+++ vista-shared/vista-domain/src/main/java/com/propertyvista/domain/policy/policies/domain/LeaseBillingTypePolicyItem.java
@@ -46,7 +46,7 @@ public interface LeaseBillingTypePolicyItem extends IEntity {
     }
 
     @OrderColumn(OrderId.class)
-    IPrimitive<Long> orderInParent();
+    IPrimitive<Integer> orderInParent();
 
     IPrimitive<BillingPeriod> billingPeriod();
